+++ yash-builtin/src/command/invoke.rs
@@ -28,6 +28,7 @@ use yash_env::semantics::Field;
 use yash_env::semantics::command::RunFunction;
 use yash_env::semantics::command::run_external_utility_in_subshell;
 use yash_env::semantics::command::search::{Target, search};
+use yash_env::system::concurrency::RunLoop;
 use yash_env::system::resource::SetRlimit;
 use yash_env::system::{
     Close, Dup, Exec, Exit, Fcntl, Fork, GetPid, IsExecutableFile, Isatty, Open, SendSignal,
@@ -48,6 +49,7 @@ impl Invoke {
             + IsExecutableFile
             + Isatty
             + Open
+            + RunLoop
             + SendSignal
             + SetPgid
             + SetRlimit
@@ -101,6 +103,7 @@ where
         + GetPid
         + Isatty
         + Open
+        + RunLoop
         + SendSignal
         + SetPgid
         + SetRlimit