+++ yash-env/src/system/virtual.rs
@@ -1217,6 +1217,62 @@ impl TcSetPgrp for VirtualSystem {
 }
 
 impl Fork for VirtualSystem {
+    /// Runs a task in a new child process.
+    ///
+    /// This method implements [`Fork::run_in_child_process`] by adding a new
+    /// process to the system state and running the given task concurrently in
+    /// the same real process. It does not create any real child process.
+    ///
+    /// To run the concurrent task, this function needs an executor that has
+    /// been set in the [`SystemState`]. If the system state does not have an
+    /// executor, this function fails with `Errno::ENOSYS`. If the executor
+    /// fails to spawn the child task, this function fails with `Errno::EAGAIN`.
+    ///
+    /// The child task must terminate itself by calling [`Exit::exit`] or
+    /// killing itself with a signal. The task returning without doing so will
+    /// cause a panic.
+    ///
+    /// The process ID of the child will be the maximum of existing process IDs
+    /// plus 1. If there are no other processes, it will be 2.
+    fn run_in_child_process<D, F>(&self, shared_data: D, child_task: F) -> (Result<Pid>, D)
+    where
+        D: Clone + 'static,
+        F: AsyncFnOnce(Self, D) + 'static,
+    {
+        let mut state = self.state.borrow_mut();
+        let Some(executor) = state.executor.as_ref() else {
+            return (Err(Errno::ENOSYS), shared_data);
+        };
+        let child_process_id = state
+            .processes
+            .keys()
+            .max()
+            .map_or(Pid(2), |pid| Pid(pid.0 + 1));
+        let child_system = VirtualSystem {
+            state: Rc::clone(&self.state),
+            process_id: child_process_id,
+        };
+        let data_clone = shared_data.clone();
+        let child_task = async move {
+            child_task(child_system.clone(), data_clone).await;
+            assert!(
+                !child_system.current_process().state().is_alive(),
+                "the child task should have terminated itself by exiting or signaling itself: pid={}",
+                child_system.process_id
+            );
+        };
+
+        if executor.spawn(Box::pin(child_task)).is_err() {
+            return (Err(Errno::EAGAIN), shared_data);
+        }
+
+        let parent_process = &state.processes[&self.process_id];
+        let child_process = Process::fork_from(self.process_id, parent_process);
+        state.processes.insert(child_process_id, child_process);
+
+        (Ok(child_process_id), shared_data)
+    }
+
     /// Creates a new child process.
     ///
     /// This implementation does not create any real child process. Instead,
@@ -2610,32 +2666,29 @@ mod tests {
     fn sigmask_resumes_after_sigcont() {
         let (system, mut executor) = virtual_system_with_executor();
         let state = Rc::clone(&system.state);
-
-        let child_process = system.new_child_process();
         let mut env = Env::with_system(system);
-        let child_process = child_process.unwrap();
-        let pid = child_process(
-            &mut env,
-            Box::new(|env| {
-                Box::pin(async move {
-                    // Block SIGTSTP (signal will become pending when raised)
-                    env.system
-                        .sigmask(Some((SigmaskOp::Add, &[SIGTSTP])), None)
-                        .await
-                        .unwrap();
-                    // Raise SIGTSTP; since it is blocked, it becomes pending
-                    env.system.raise(SIGTSTP).await.unwrap();
-                    // Unblocking SIGTSTP delivers it, stopping the process; the
-                    // future suspends until the process is resumed
-                    env.system
-                        .sigmask(Some((SigmaskOp::Remove, &[SIGTSTP])), None)
-                        .await
-                        .unwrap();
-                    // After being resumed, exit cleanly
-                    env.system.exit(ExitStatus(0)).await
-                })
-            }),
-        );
+        let pid = env
+            .run_in_child_process((), |child_env: Env<VirtualSystem>, ()| async move {
+                // Block SIGTSTP (signal will become pending when raised)
+                child_env
+                    .system
+                    .sigmask(Some((SigmaskOp::Add, &[SIGTSTP])), None)
+                    .await
+                    .unwrap();
+                // Raise SIGTSTP; since it is blocked, it becomes pending
+                child_env.system.raise(SIGTSTP).await.unwrap();
+                // Unblocking SIGTSTP delivers it, stopping the process; the
+                // future suspends until the process is resumed
+                child_env
+                    .system
+                    .sigmask(Some((SigmaskOp::Remove, &[SIGTSTP])), None)
+                    .await
+                    .unwrap();
+                // After being resumed, exit cleanly
+                child_env.system.exit(ExitStatus(0)).await;
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         // The child should now be stopped
@@ -2869,8 +2922,7 @@ mod tests {
         let (system, _executor) = virtual_system_with_executor();
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        let child = env.system.new_child_process().unwrap();
-        let pid = child(&mut env, Box::new(|_env| Box::pin(pending())));
+        let pid = env.run_in_child_process((), |_, ()| pending()).0.unwrap();
 
         let result = env.system.setpgid(pid, pid);
         assert_eq!(result, Ok(()));
@@ -2884,17 +2936,14 @@ mod tests {
         let (system, mut executor) = virtual_system_with_executor();
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        let child = env.system.new_child_process().unwrap();
-        let pid = child(
-            &mut env,
-            Box::new(|child_env| {
-                Box::pin(async move {
-                    let result = child_env.system.setpgid(Pid(0), Pid(0));
-                    assert_eq!(result, Ok(()));
-                    child_env.system.exit(child_env.exit_status).await
-                })
-            }),
-        );
+        let pid = env
+            .run_in_child_process((), |child_env: Env<VirtualSystem>, ()| async move {
+                let result = child_env.system.setpgid(Pid(0), Pid(0));
+                assert_eq!(result, Ok(()));
+                child_env.system.exit(child_env.exit_status).await;
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         let pgid = state.borrow().processes[&pid].pgid();
@@ -2906,11 +2955,9 @@ mod tests {
         let (system, _executor) = virtual_system_with_executor();
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        let child_1 = env.system.new_child_process().unwrap();
-        let pid_1 = child_1(&mut env, Box::new(|_env| Box::pin(pending())));
+        let pid_1 = env.run_in_child_process((), |_, ()| pending()).0.unwrap();
         env.system.setpgid(pid_1, pid_1).unwrap();
-        let child_2 = env.system.new_child_process().unwrap();
-        let pid_2 = child_2(&mut env, Box::new(|_env| Box::pin(pending())));
+        let pid_2 = env.run_in_child_process((), |_, ()| pending()).0.unwrap();
 
         let result = env.system.setpgid(pid_2, pid_1);
         assert_eq!(result, Ok(()));
@@ -2924,8 +2971,7 @@ mod tests {
         let (system, _executor) = virtual_system_with_executor();
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        let child = env.system.new_child_process().unwrap();
-        let pid = child(&mut env, Box::new(|_env| Box::pin(pending())));
+        let pid = env.run_in_child_process((), |_, ()| pending()).0.unwrap();
 
         let dummy_pid = Pid(123);
         let result = env.system.setpgid(dummy_pid, dummy_pid);
@@ -2941,17 +2987,14 @@ mod tests {
         let parent_pid = system.process_id;
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        let child = env.system.new_child_process().unwrap();
-        let _pid = child(
-            &mut env,
-            Box::new(move |child_env| {
-                Box::pin(async move {
-                    let result = child_env.system.setpgid(parent_pid, Pid(0));
-                    assert_eq!(result, Err(Errno::ESRCH));
-                    child_env.system.exit(child_env.exit_status).await
-                })
-            }),
-        );
+        let _pid = env
+            .run_in_child_process((), move |child_env: Env<VirtualSystem>, ()| async move {
+                let result = child_env.system.setpgid(parent_pid, Pid(0));
+                assert_eq!(result, Err(Errno::ESRCH));
+                child_env.system.exit(child_env.exit_status).await;
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         let pgid = state.borrow().processes[&parent_pid].pgid();
@@ -2972,17 +3015,14 @@ mod tests {
         let state = Rc::clone(&system.state);
         state.borrow_mut().file_system.save(path, content).unwrap();
         let mut env = Env::with_system(system);
-        let child = env.system.new_child_process().unwrap();
-        let pid = child(
-            &mut env,
-            Box::new(move |child_env| {
-                Box::pin(async move {
-                    let path = CString::new(path).unwrap();
-                    child_env.system.execve(&path, &[], &[]).await.ok();
-                    child_env.system.exit(child_env.exit_status).await
-                })
-            }),
-        );
+        let pid = env
+            .run_in_child_process((), move |child_env: Env<VirtualSystem>, ()| async move {
+                let path = CString::new(path).unwrap();
+                child_env.system.execve(&path, &[], &[]).await.ok();
+                child_env.system.exit(child_env.exit_status).await;
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         let result = env.system.setpgid(pid, pid);
@@ -2997,11 +3037,9 @@ mod tests {
         let (system, mut executor) = virtual_system_with_executor();
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        let child_1 = env.system.new_child_process().unwrap();
-        let pid_1 = child_1(&mut env, Box::new(|_env| Box::pin(pending())));
+        let pid_1 = env.run_in_child_process((), |_, ()| pending()).0.unwrap();
         // env.system.setpgid(pid_1, pid_1).unwrap();
-        let child_2 = env.system.new_child_process().unwrap();
-        let pid_2 = child_2(&mut env, Box::new(|_env| Box::pin(pending())));
+        let pid_2 = env.run_in_child_process((), |_, ()| pending()).0.unwrap();
         executor.run_until_stalled();
 
         let result = env.system.setpgid(pid_2, pid_1);
@@ -3051,8 +3089,86 @@ mod tests {
     }
 
     #[test]
+    fn run_in_child_process_shares_data() {
+        let (system, mut executor) = virtual_system_with_executor();
+        let data = "foo".to_string();
+        let (_pid_or_error, data) =
+            system.run_in_child_process(data, |child_system: VirtualSystem, data| async move {
+                assert_eq!(data, "foo");
+                child_system.exit(ExitStatus(0)).await;
+            });
+
+        assert_eq!(data, "foo");
+        executor.run_until_stalled();
+    }
+
+    #[test]
+    fn run_in_child_process_chooses_minimal_unused_pid() {
+        let (system, _executor) = virtual_system_with_executor();
+        let dummy_process = Process::with_parent_and_group(system.process_id, Pid(1));
+        system
+            .state
+            .borrow_mut()
+            .processes
+            .insert(Pid(42), dummy_process);
+
+        let (pid_or_error, ()) = system.run_in_child_process((), |_, ()| pending());
+        assert_eq!(pid_or_error, Ok(Pid(43)));
+    }
+
+    #[test]
+    fn run_in_child_process_adds_child_to_process_table() {
+        let (system, mut executor) = virtual_system_with_executor();
+        let child_pid = Rc::new(Cell::new(None));
+
+        let (pid_or_error, child_pid) = system.run_in_child_process(
+            child_pid,
+            |child_system: VirtualSystem, child_pid: Rc<Cell<Option<Pid>>>| async move {
+                child_pid.set(Some(child_system.getpid()));
+                child_system.exit(ExitStatus(0)).await;
+            },
+        );
+        executor.run_until_stalled();
+
+        let child_pid = child_pid.get().expect("child task should have set pid");
+        assert_eq!(pid_or_error, Ok(child_pid));
+        let state = system.state.borrow();
+        let child_process = state
+            .processes
+            .get(&child_pid)
+            .expect("child process should be in process table");
+        assert_eq!(child_process.ppid(), system.process_id);
+    }
+
+    #[test]
+    #[should_panic = "the child task should have terminated itself by exiting or signaling itself: pid=3"]
+    fn run_in_child_process_panics_on_child_task_not_exiting() {
+        let (system, mut executor) = virtual_system_with_executor();
+
+        system
+            .run_in_child_process((), |_, ()| async { /* exit() */ })
+            .0
+            .unwrap();
+
+        executor.run_until_stalled();
+    }
+
+    #[test]
+    fn run_in_child_process_fails_without_executor() {
+        let system = VirtualSystem::new();
+        let result = system
+            .run_in_child_process((), |_, ()| {
+                unreachable!("child task should not run without executor")
+                    as std::future::Pending<()>
+            })
+            .0;
+        assert_eq!(result, Err(Errno::ENOSYS));
+    }
+
+    #[test]
     fn new_child_process_without_executor() {
         let system = VirtualSystem::new();
+        #[allow(deprecated)]
         let result = system.new_child_process();
         match result {
             Ok(_) => panic!("unexpected Ok value"),
@@ -3064,6 +3180,7 @@ mod tests {
     fn new_child_process_with_executor() {
         let (system, _executor) = virtual_system_with_executor();
 
+        #[allow(deprecated)]
         let result = system.new_child_process();
 
         let state = system.state.borrow();
@@ -3082,19 +3199,14 @@ mod tests {
     #[test]
     fn wait_for_running_child() {
         let (system, _executor) = virtual_system_with_executor();
-
-        let child_process = system.new_child_process();
-
         let mut env = Env::with_system(system);
-        let child_process = child_process.unwrap();
-        let pid = child_process(
-            &mut env,
-            Box::new(|_env| {
-                Box::pin(async {
-                    unreachable!("child process does not progress unless executor is used")
-                })
-            }),
-        );
+
+        let pid = env
+            .run_in_child_process((), |_, ()| async {
+                unreachable!("child process does not progress unless executor is used")
+            })
+            .0
+            .unwrap();
 
         let result = env.system.wait(pid);
         assert_eq!(result, Ok(None))
@@ -3103,15 +3215,14 @@ mod tests {
     #[test]
     fn wait_for_exited_child() {
         let (system, mut executor) = virtual_system_with_executor();
-
-        let child_process = system.new_child_process();
-
         let mut env = Env::with_system(system);
-        let child_process = child_process.unwrap();
-        let pid = child_process(
-            &mut env,
-            Box::new(|env| Box::pin(async move { env.system.exit(ExitStatus(5)).await })),
-        );
+
+        let pid = env
+            .run_in_child_process((), |child_env: Env<VirtualSystem>, ()| async move {
+                child_env.system.exit(ExitStatus(5)).await;
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         let result = env.system.wait(pid);
@@ -3121,21 +3232,16 @@ mod tests {
     #[test]
     fn wait_for_signaled_child() {
         let (system, mut executor) = virtual_system_with_executor();
-
-        let child_process = system.new_child_process();
-
         let mut env = Env::with_system(system);
-        let child_process = child_process.unwrap();
-        let pid = child_process(
-            &mut env,
-            Box::new(|env| {
-                Box::pin(async move {
-                    let pid = env.system.getpid();
-                    let result = env.system.kill(pid, Some(SIGKILL)).await;
-                    unreachable!("kill returned {result:?}");
-                })
-            }),
-        );
+
+        let pid = env
+            .run_in_child_process((), |child_env: Env<VirtualSystem>, ()| async move {
+                let pid = child_env.system.getpid();
+                let result = child_env.system.kill(pid, Some(SIGKILL)).await;
+                unreachable!("kill returned {result:?}");
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         let result = env.system.wait(pid);
@@ -3154,21 +3260,16 @@ mod tests {
     #[test]
     fn wait_for_stopped_child() {
         let (system, mut executor) = virtual_system_with_executor();
-
-        let child_process = system.new_child_process();
-
         let mut env = Env::with_system(system);
-        let child_process = child_process.unwrap();
-        let pid = child_process(
-            &mut env,
-            Box::new(|env| {
-                Box::pin(async move {
-                    let pid = env.system.getpid();
-                    let result = env.system.kill(pid, Some(SIGSTOP)).await;
-                    unreachable!("kill returned {result:?}");
-                })
-            }),
-        );
+
+        let pid = env
+            .run_in_child_process((), |child_env: Env<VirtualSystem>, ()| async move {
+                let pid = child_env.system.getpid();
+                let result = child_env.system.kill(pid, Some(SIGSTOP)).await;
+                unreachable!("kill returned {result:?}");
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         let result = env.system.wait(pid);
@@ -3178,22 +3279,16 @@ mod tests {
     #[test]
     fn wait_for_resumed_child() {
         let (system, mut executor) = virtual_system_with_executor();
-
-        let child_process = system.new_child_process();
-
         let mut env = Env::with_system(system);
-        let child_process = child_process.unwrap();
-        let pid = child_process(
-            &mut env,
-            Box::new(|env| {
-                Box::pin(async move {
-                    let pid = env.system.getpid();
-                    let result = env.system.kill(pid, Some(SIGSTOP)).await;
-                    assert_eq!(result, Ok(()));
-                    env.system.exit(ExitStatus(123)).await
-                })
-            }),
-        );
+
+        let pid = env
+            .run_in_child_process((), |child_env: Env<VirtualSystem>, ()| async move {
+                let pid = child_env.system.getpid();
+                child_env.system.kill(pid, Some(SIGSTOP)).await.unwrap();
+                child_env.system.exit(ExitStatus(123)).await;
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         env.system
@@ -3226,23 +3321,6 @@ mod tests {
         // TODO
         // let result = system.wait(Pid(-1234));
         // assert_eq!(result, Err(Errno::ECHILD));
-    }
-
-    #[test]
-    fn exiting_child_sends_sigchld_to_parent() {
-        let (system, mut executor) = virtual_system_with_executor();
-        system.sigaction(SIGCHLD, Disposition::Catch).unwrap();
-
-        let child_process = system.new_child_process().unwrap();
-        let system2 = system.clone();
-        let mut env = Env::with_system(system);
-        let _pid = child_process(
-            &mut env,
-            Box::new(|env| Box::pin(async { env.system.exit(ExitStatus(0)).await })),
-        );
-        executor.run_until_stalled();
-
-        assert_eq!(system2.current_process().caught_signals, [SIGCHLD]);
     }
 
     #[test]
@@ -3331,14 +3409,15 @@ mod tests {
     fn exit_sends_sigchld_to_parent() {
         let (system, mut executor) = virtual_system_with_executor();
         system.sigaction(SIGCHLD, Disposition::Catch).unwrap();
-
-        let child_process = system.new_child_process().unwrap();
         let system2 = system.clone();
         let mut env = Env::with_system(system);
-        let _pid = child_process(
-            &mut env,
-            Box::new(|env| Box::pin(async { env.system.exit(ExitStatus(123)).await })),
-        );
+
+        let _pid = env
+            .run_in_child_process((), |child_env: Env<VirtualSystem>, ()| async move {
+                child_env.system.exit(ExitStatus(123)).await;
+            })
+            .0
+            .unwrap();
         executor.run_until_stalled();
 
         assert_eq!(system2.current_process().caught_signals, [SIGCHLD]);